# src/main/resources/META-INF/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: support@wefox.com
    name: Wefox
    url: https://wefox.com
  description: |-
    API exposed by the new GEO Service, which allows to query for cities given (country/region) or (country/region/province).

    The IDs shown in the examples are examples. For simplicity some examples show subsequent IDs, but IDs should not be autoincremental, for security reasons it's better to use UUID instead.

    **The names of countries, regions and cities returned by all endpoints are all lowercase**. It's up to users to format the name: capitalize first letter or uppercase.
  title: geo-service
  version: "1.0"
servers:
- description: Local
  url: http://localhost:8080/geo
security:
- API Key - 1: []
paths:
  /countries:
    get:
      description: |-
        Get all countries. The names are resolved based on requested language (default `it`).

        This API is paginated. The parameters `limit` and `offset` could be omitted but default values will be used. The 'links' section includes the link to navigate the collection.

        See [HATEOAS](https://restfulapi.net/hateoas/) for further informations about `links` section.
      operationId: get-countries
      parameters:
      - allowEmptyValue: true
        description: the language in which we want to receive the data, expressed
          in iso 639-1 format
        explode: true
        in: query
        name: lang
        required: false
        schema:
          default: it
          maxLength: 2
          pattern: '[A-Za-z]{2}'
          type: string
        style: form
      - allowEmptyValue: true
        description: the offset to start
        explode: true
        in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          type: number
        style: form
      - allowEmptyValue: true
        description: the page size, for collection APIs
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 20
          maximum: 50
          minimum: 0
          type: number
        style: form
      responses:
        "200":
          content:
            application/json:
              examples:
                Example 1 (api called with lang=en):
                  value:
                    data:
                    - country_id: it
                      name: italy
                      phone_code: "39"
                      deleted: false
                    - country_id: es
                      name: spain
                      phone_code: "34"
                      deleted: false
                    - country_id: de
                      name: germany
                      phone_code: "49"
                      deleted: false
                    links:
                    - href: /geo/countries?offset=0&limit=20&lang=en
                      rel: first
                    - href: /geo/countries?offset=20&limit=20&lang=en
                      rel: prev
                    - href: /geo/countries?offset=60&limit=20&lang=en
                      rel: next
                    - href: /geo/countries?offset=230&limit=20&lang=en
                      rel: last
                    metadata:
                      current_offset: 40
                      total_count: 239
                Example 2 (api called with lang=it):
                  value:
                    data:
                    - country_id: it
                      name: italia
                      phone_code: "39"
                      deleted: false
                    - country_id: es
                      name: spagna
                      deleted: false
                      phone_code: "34"
                    - country_id: de
                      name: germania
                      deleted: false
                      phone_code: "49"
                    links:
                    - href: /geo/countries?offset=0&limit=20&lang=it
                      rel: first
                    - href: /geo/countries?offset=20&limit=20&lang=it
                      rel: prev
                    - href: /geo/countries?offset=60&limit=20&lang=it
                      rel: next
                    - null
                    - href: /geo/countries?offset=230&limit=20&lang=it
                      rel: last
                    metadata:
                      current_offset: 40
                      total_count: 239
              schema:
                $ref: '#/components/schemas/inline_response_200'
          description: |-
            List of countries, ordered by name. The examples show a minimal subset of the countries just to give an idea. Let's suppose that in each result we'll find a number of results equals to `limit` parameter.

            See [HATEOAS](https://restfulapi.net/hateoas/) for further informations about `links` section.

            The examples shows a subset of the results. Example 1 shows the result when called with `lang=en`, while Example 2 shows the list with `lang=it`.
          headers:
            X-RateLimit-Limit:
              explode: false
              schema:
                maxLength: 512
                pattern: ^[A-Za-z0-9-_]+
                type: string
              style: simple
        "401":
          content:
            application/json:
              examples:
                HTTP 401 body:
                  value:
                    errors:
                    - code: ERR001
                      message: Some error description
              schema:
                $ref: '#/components/schemas/Errors'
          description: Unauthorized
          headers:
            X-RateLimit-Limit:
              explode: false
              schema:
                maxLength: 512
                pattern: ^[A-Za-z0-9-_]+
                type: string
              style: simple
        "403":
          content:
            application/json:
              examples:
                HTTP 403 body:
                  value:
                    errors:
                    - code: ERR403
                      message: Some error
              schema:
                $ref: '#/components/schemas/Errors'
          description: Forbidden access
          headers:
            X-RateLimit-Limit:
              explode: false
              schema:
                maxLength: 512
                pattern: ^[A-Za-z0-9-_]+
                type: string
              style: simple
        "500":
          content:
            application/json:
              examples:
                HTTP 500 body:
                  value:
                    errors:
                    - code: ERR500
                      message: Some error
              schema:
                $ref: '#/components/schemas/Errors'
          description: Error response
          headers:
            X-RateLimit-Limit:
              explode: false
              schema:
                maxLength: 512
                pattern: ^[A-Za-z0-9-_]+
                type: string
              style: simple
      summary: All Countries
      tags: []
      x-accepts: application/json
      x-tags: []
...
  /countries/{country_id}/regions/{region_id}/cities:
    get:
      description: |-
        Returns the cities of a region. The names are resolved based on requested language (default `it`).

        The `links` section includes the link to navigate the collection.

        See [HATEOAS](https://restfulapi.net/hateoas/) for further informations about `links` section.
      operationId: cities-by-region
      parameters:
      - explode: false
        in: path
        name: country_id
        required: true
        schema:
          maxLength: 2
          pattern: '[A-Za-z]{2}'
          type: string
        style: simple
      - explode: false
        in: path
        name: region_id
        required: true
        schema:
          maxLength: 10
          pattern: '[A-Za-z0-9]+'
          type: string
        style: simple
      - allowEmptyValue: true
        description: the language in which we want to receive the data, expressed
          in iso 639-1 format
        explode: true
        in: query
        name: lang
        required: false
        schema:
          default: it
          maxLength: 2
          pattern: '[A-Za-z]{2}'
          type: string
        style: form
      - allowEmptyValue: true
        description: the page size, for collection APIs
        explode: true
        in: query
        name: limit
        required: false
        schema:
          default: 20
          maximum: 50
          minimum: 0
          type: number
        style: form
      - allowEmptyValue: true
        description: the offset to start
        explode: true
        in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          type: number
        style: form
      responses:
        "200":
          content:
            application/json:
              examples:
                Example 1 (cities of Lombardy region, Italy):
                  value:
                    data:
                    - city_id: itc9103
                      name: abbadia cerreto
                      deleted: false
                      state_code: "98001"
                      management_code: A004
                      zip_codes:
                      - "26834"
                      region_id: it-lo
                      region_name: lombardia
                      province_id: itp002
                      province_name: lodi
                      country_id: it
                      country_name: italia
                    - city_id: itc10601
                      name: abbadia lariana
                      deleted: false
                      state_code: "97001"
                      management_code: A005
                      zip_codes:
                      - "23821"
                      region_id: it-lo
                      region_name: lombardia
                      province_id: itp006
                      province_name: como
                      country_id: it
                      country_name: italia
                    links:
                    - href: /geo/countries/it/regions/it-lo/cities?offset=0&limit=20&lang=it
                      rel: first
                    - href: /geo/countries/it/regions/it-lo/cities?offset=20&limit=20&lang=it
                      rel: prev
                    - href: /geo/countries/it/regions/it-lo/cities?offset=60&limit=20&lang=it
                      rel: next
                    - href: /geo/countries/it/regions/it-lo/cities?offset=1500&limit=20&lang=it
                      rel: last
                    metadata:
                      current_offset: 40
                      total_count: 1506
                Example 2 (cities of the province of Milan, in Lombardy):
                  value:
                    data:
                    - city_id: itc10104
                      name: abbiategrasso
                      deleted: false
                      state_code: "15002"
                      management_code: A010
                      zip_codes:
                      - "20081"
                      region_id: it-lo
                      region_name: lombardia
                      province_id: itp001
                      province_name: milano
                      country_id: it
                      country_name: italia
                    - city_id: itc10105
                      name: Albairate
                      deleted: false
                      state_code: "15005"
                      management_code: A127
                      zip_codes:
                      - "20080"
                      region_id: it-lo
                      region_name: lombardia
                      province_id: itp001
                      province_name: milano
                      country_id: it
                      country_name: italia
                    links:
                    - href: /geo/countries/it/regions/it-lo/provinces/itp001/cities?offset=0&limit=20&lang=it
                      rel: first
                    - href: /geo/countries/it/regions/it-lo/provinces/itp001/cities?offset=20&limit=20&lang=it
                      rel: prev
                    - href: /geo/countries/it/regions/it-lo/provinces/itp001/cities?offset=60&limit=20&lang=it
                      rel: next
                    - href: /geo/countries/it/regions/it-lo/provinces/itp001/cities?offset=120&limit=20&lang=it
                      rel: last
                    metadata:
                      current_offset: 40
                      total_count: 134
              schema:
                $ref: '#/components/schemas/inline_response_200_2'
          description: |-
            List of cities. This API can be called filtering by province of a region of by a region. The result is ordered by name, according to the `lang` param.

            Notice that the examples shows a little subset of data, not the whole set, so consider it as an example. In particular it's a little set of the cities of Lombardy (Example1) and of Milan province (a district of Lombardy, Example 2).

            See [HATEOAS](https://restfulapi.net/hateoas/) for further informations about `links` section.
          headers:
            X-RateLimit-Limit:
              explode: false
              schema:
                maxLength: 512
                pattern: ^[A-Za-z0-9-_]+
                type: string
              style: simple
        "400":
          content:
            application/json:
              examples:
                HTTP 400 body:
                  value:
                    errors:
                    - code: ERR001
                      message: An error message
              schema:
                $ref: '#/components/schemas/Errors'
          description: Invalid request
          headers:
            X-RateLimit-Limit:
              explode: false
              schema:
                maxLength: 512
                pattern: ^[A-Za-z0-9-_]+
                type: string
              style: simple
        "401":
          content:
            application/json:
              examples:
                HTTP 401 body:
                  value:
                    errors:
                    - code: ERR001
                      message: Some error description
              schema:
                $ref: '#/components/schemas/Errors'
          description: Unauthorized
          headers:
            X-RateLimit-Limit:
              explode: false
              schema:
                maxLength: 512
                pattern: ^[A-Za-z0-9-_]+
                type: string
              style: simple
        "403":
          content:
            application/json:
              examples:
                HTTP 403 body:
                  value:
                    errors:
                    - code: ERR403
                      message: Some error
              schema:
                $ref: '#/components/schemas/Errors'
          description: Forbidden access
          headers:
            X-RateLimit-Limit:
              explode: false
              schema:
                maxLength: 512
                pattern: ^[A-Za-z0-9-_]+
                type: string
              style: simple
        "404":
          content:
            application/json:
              examples:
                HTTP 404 body:
                  value:
                    errors:
                    - code: ERR404
                      message: Some error
              schema:
                $ref: '#/components/schemas/Errors'
          description: Not found
          headers:
            X-RateLimit-Limit:
              explode: false
              schema:
                maxLength: 512
                pattern: ^[A-Za-z0-9-_]+
                type: string
              style: simple
        "500":
          content:
            application/json:
              examples:
                HTTP 500 body:
                  value:
                    errors:
                    - code: ERR500
                      message: Some error
              schema:
                $ref: '#/components/schemas/Errors'
          description: Error response
          headers:
            X-RateLimit-Limit:
              explode: false
              schema:
                maxLength: 512
                pattern: ^[A-Za-z0-9-_]+
                type: string
              style: simple
      summary: Cities by region
      tags: []
      x-accepts: application/json
      x-tags: []
  /countries/{country_id}/regions/{region_id}/cities/{city_id}:
    get:
      deprecated: true
      description: Returns a specific city of a region. The names are resolved based
        on requested language (default `it`)
      operationId: cities-by-region-and-city-id
      parameters:
      - explode: false
        in: path
        name: country_id
        required: true
        schema:
          maxLength: 2
          pattern: '[A-Za-z]{2}'
          type: string
        style: simple
      - explode: false
        in: path
        name: region_id
        required: true
        schema:
          maxLength: 10
          pattern: '[A-Za-z0-9]+'
          type: string
        style: simple
      - explode: false
        in: path
        name: city_id
        required: true
        schema:
          maxLength: 10
          pattern: '[A-Za-z0-9]+'
          type: string
        style: simple
      - allowEmptyValue: true
        description: the language in which we want to receive the data, expressed
          in iso 639-1 format
        explode: true
        in: query
        name: lang
        required: false
        schema:
          default: it
          maxLength: 2
          pattern: '[A-Za-z]{2}'
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              examples:
                Detail of a city (lang=it):
                  value:
                    data:
                      city_id: itc10104
                      name: linarolo
                      state_code: "18081"
                      management_code: E600
                      zip_codes:
                      - "27010"
                      country_id: it
                      country_name: italia
                      region_id: it-lo
                      region_name: lombardia
                      province_id: itp008
                      province_name: Pavia
              schema:
                $ref: '#/components/schemas/inline_response_200_3'
          description: Detail of single city. Region, Province (if available) and
            Country are resolved
          headers:
            X-RateLimit-Limit:
              explode: false
              schema:
                maxLength: 512
                pattern: ^[A-Za-z0-9-_]+
                type: string
              style: simple
        "400":
          content:
            application/json:
              examples:
                HTTP 400 body:
                  value:
                    errors:
                    - code: ERR001
                      message: An error message
              schema:
                $ref: '#/components/schemas/Errors'
          description: Invalid request
          headers:
            X-RateLimit-Limit:
              explode: false
              schema:
                maxLength: 512
                pattern: ^[A-Za-z0-9-_]+
                type: string
              style: simple
        "401":
          content:
            application/json:
              examples:
                HTTP 401 body:
                  value:
                    errors:
                    - code: ERR001
                      message: Some error description
              schema:
                $ref: '#/components/schemas/Errors'
          description: Unauthorized
          headers:
            X-RateLimit-Limit:
              explode: false
              schema:
                maxLength: 512
                pattern: ^[A-Za-z0-9-_]+
                type: string
              style: simple
        "403":
          content:
            application/json:
              examples:
                HTTP 403 body:
                  value:
                    errors:
                    - code: ERR403
                      message: Some error
              schema:
                $ref: '#/components/schemas/Errors'
          description: Forbidden access
          headers:
            X-RateLimit-Limit:
              explode: false
              schema:
                maxLength: 512
                pattern: ^[A-Za-z0-9-_]+
                type: string
              style: simple
        "404":
          content:
            application/json:
              examples:
                HTTP 404 body:
                  value:
                    errors:
                    - code: ERR404
                      message: Some error
              schema:
                $ref: '#/components/schemas/Errors'
          description: Not found
          headers:
            X-RateLimit-Limit:
              explode: false
              schema:
                maxLength: 512
                pattern: ^[A-Za-z0-9-_]+
                type: string
              style: simple
        "500":
          content:
            application/json:
              examples:
                HTTP 500 body:
                  value:
                    errors:
                    - code: ERR500
                      message: Some error
              schema:
                $ref: '#/components/schemas/Errors'
          description: Error response
          headers:
            X-RateLimit-Limit:
              explode: false
              schema:
                maxLength: 512
                pattern: ^[A-Za-z0-9-_]+
                type: string
              style: simple
      summary: City detail by city ID and region
      tags: []
      x-accepts: application/json
      x-tags: []
...
components:
...
  schemas:
...
    Link:
      description: '  HATEOAS link'
      example:
        rel: rel
        href: https://openapi-generator.tech
      properties:
        href:
          format: uri
          maxLength: 512
          type: string
        rel:
          maxLength: 20
          pattern: '[A-Za-z0-9 ]+'
          type: string
      title: Link
      type: object
      x-stoplight:
        id: k4ckz809rljnf
...
    Errors:
      description: A list of errors
      properties:
        errors:
          items:
            $ref: '#/components/schemas/Error'
          maxItems: 20
          type: array
      title: Errors
      type: object
      x-stoplight:
        id: 42dh0daenq6h2
    inline_response_200:
      example:
        metadata:
          current_offset: 8008
          total_count: 6027
        data:
        - deleted: false
          name: name
          country_id: country_id
          phone_code: phone_code
        - deleted: false
          name: name
          country_id: country_id
          phone_code: phone_code
        - deleted: false
          name: name
          country_id: country_id
          phone_code: phone_code
        - deleted: false
          name: name
          country_id: country_id
          phone_code: phone_code
        - deleted: false
          name: name
          country_id: country_id
          phone_code: phone_code
        links:
        - rel: rel
          href: https://openapi-generator.tech
        - rel: rel
          href: https://openapi-generator.tech
        - rel: rel
          href: https://openapi-generator.tech
        - rel: rel
          href: https://openapi-generator.tech
        - rel: rel
          href: https://openapi-generator.tech
      properties:
        data:
          items:
            $ref: '#/components/schemas/Country'
          maxItems: 300
          type: array
        links:
          items:
            $ref: '#/components/schemas/Link'
          maxItems: 20
          type: array
        metadata:
          $ref: '#/components/schemas/Metadata'
      required:
      - data
      type: object
...
    inline_response_200_2:
      example:
        metadata:
          current_offset: 8008
          total_count: 6027
        data:
        - inhabitants: 11201
          extension: 180.07
          region_id: region_id
          province_name: province_name
          deleted: false
          province_id: province_id
          name: name
          country_name: country_name
          region_name: region_name
          management_code: management_code
          state_code: state_code
          zip_codes:
          - zip_codes
          - zip_codes
          - zip_codes
          - zip_codes
          - zip_codes
          country_id: country_id
          city_id: city_id
        - inhabitants: 11201
          extension: 180.07
          region_id: region_id
          province_name: province_name
          deleted: false
          province_id: province_id
          name: name
          country_name: country_name
          region_name: region_name
          management_code: management_code
          state_code: state_code
          zip_codes:
          - zip_codes
          - zip_codes
          - zip_codes
          - zip_codes
          - zip_codes
          country_id: country_id
          city_id: city_id
        - inhabitants: 11201
          extension: 180.07
          region_id: region_id
          province_name: province_name
          deleted: false
          province_id: province_id
          name: name
          country_name: country_name
          region_name: region_name
          management_code: management_code
          state_code: state_code
          zip_codes:
          - zip_codes
          - zip_codes
          - zip_codes
          - zip_codes
          - zip_codes
          country_id: country_id
          city_id: city_id
        - inhabitants: 11201
          extension: 180.07
          region_id: region_id
          province_name: province_name
          deleted: false
          province_id: province_id
          name: name
          country_name: country_name
          region_name: region_name
          management_code: management_code
          state_code: state_code
          zip_codes:
          - zip_codes
          - zip_codes
          - zip_codes
          - zip_codes
          - zip_codes
          country_id: country_id
          city_id: city_id
        - inhabitants: 11201
          extension: 180.07
          region_id: region_id
          province_name: province_name
          deleted: false
          province_id: province_id
          name: name
          country_name: country_name
          region_name: region_name
          management_code: management_code
          state_code: state_code
          zip_codes:
          - zip_codes
          - zip_codes
          - zip_codes
          - zip_codes
          - zip_codes
          country_id: country_id
          city_id: city_id
        links:
        - rel: rel
          href: https://openapi-generator.tech
        - rel: rel
          href: https://openapi-generator.tech
        - rel: rel
          href: https://openapi-generator.tech
        - rel: rel
          href: https://openapi-generator.tech
        - rel: rel
          href: https://openapi-generator.tech
      properties:
        data:
          items:
            $ref: '#/components/schemas/City'
          maxItems: 1000
          type: array
        links:
          items:
            $ref: '#/components/schemas/Link'
          maxItems: 20
          type: array
        metadata:
          $ref: '#/components/schemas/Metadata'
      required:
      - data
      type: object
    inline_response_200_3:
      example:
        data:
          inhabitants: 11201
          extension: 180.07
          region_id: region_id
          province_name: province_name
          deleted: false
          province_id: province_id
          name: name
          country_name: country_name
          region_name: region_name
          management_code: management_code
          state_code: state_code
          zip_codes:
          - zip_codes
          - zip_codes
          - zip_codes
          - zip_codes
          - zip_codes
          country_id: country_id
          city_id: city_id
      properties:
        data:
          $ref: '#/components/schemas/City'
      required:
      - data
      type: object
...
  securitySchemes:
    x-geoservice-key:
      in: header
      name: api-key
      type: apiKey
...